CVS-95440 C api start impl (#1509)
First part of server start implementation:
* separate paths for OVMS server start and C-API server start
* minimum implementation for --config_path --port and --rest_port
* dummy implementation for the rest
* mini demo app demonstrating server start via C-API
* MakefileCapi modified to run standalone mini demo app outside of OVMS
* increase coverage for config death tests

+++ src/modelconfig.cpp
@@ -537,7 +537,7 @@ Status ModelConfig::parseNode(const rapidjson::Value& v) {
                     for (auto& sh : s.value.GetArray()) {
                         if (sh.IsUint64()) {
                             size_t dim = sh.GetUint64();
-                            if (dim > std::numeric_limits<dimension_value_t>::max()) {
+                            if (dim > static_cast<size_t>(std::numeric_limits<dimension_value_t>::max())) {
                                 SPDLOG_WARN("There was an error parsing shape {}", s.name.GetString());
                                 if (!firstErrorStatus.ok()) {
                                     firstErrorStatus = StatusCode::SHAPE_WRONG_FORMAT;
